Using logging objects in API player
