+++ core_helpers/logging_objects.py
@@ -4,13 +4,21 @@ Helper class for Logging Objects
 
 from utils.Base_Logging import Base_Logging
 from utils.stop_test_exception_util import Stop_Test_Exception
+from utils import Gif_Maker
 import logging
 
-class Logging_Objects:
-    def __init__(self):
+class Logging_Objects():
+    def __init__(self, base_url=None):
         self.msg_list = []
         self.exceptions = []
         self.browserstack_flag = False
+        self.mini_check_counter = 0
+        self.mini_check_pass_counter = 0
+        self.result_counter = 0
+        self.pass_counter = 0
+        self.failure_message_list = []
+        self.screenshot_dir = None
+        self.calling_module = None
 
     def write_test_summary(self):
         "Print out a useful, human readable summary"
@@ -29,11 +37,10 @@ class Logging_Objects:
             for (i,msg) in enumerate(self.exceptions,start=1):
                 self.write(str(i)+"- " + msg)
 
-        if self.api_test_flag is False:
-            self.make_gif()
-            if self.gif_file_name is not None:
-                self.write("Screenshots & GIF created at %s"%self.screenshot_dir)
-                self.write('************************')
+        self.make_gif()
+        if self.gif_file_name is not None:
+            self.write("Screenshots & GIF created at %s"%self.screenshot_dir)
+            self.write('************************')
 
     def write(self,msg,level='info', trace_back=None):
         "Log the message"
@@ -88,3 +95,49 @@ class Logging_Objects:
     def set_rp_logger(self,rp_pytest_service):
         "Set the reportportal logger"
         self.rp_logger = self.log_obj.setup_rp_logging(rp_pytest_service)
+
+    def conditional_write(self,flag,positive,negative,level='info'):
+        "Write out either the positive or the negative message based on flag"
+        self.mini_check_counter += 1
+        if level.lower() == "inverse":
+            if flag is True:
+                self.write(positive,level='error')
+            else:
+                self.write(negative,level='info')
+                self.mini_check_pass_counter += 1
+        else:
+            if flag is True:
+                self.write(positive,level='info')
+                self.mini_check_pass_counter += 1
+            else:
+                self.write(negative,level='error')
+
+    def make_gif(self):
+        "Create a gif of all the screenshots within the screenshots directory"
+        self.gif_file_name = Gif_Maker.make_gif(self.screenshot_dir,name=self.calling_module)
+
+        return self.gif_file_name
+
+    def set_calling_module(self,name):
+        "Set the test name"
+        self.calling_module = name
+
+    def get_calling_module(self):
+        "Get the name of the calling module"
+        if self.calling_module is None:
+            #Try to intelligently figure out name of test when not using pytest
+            full_stack = inspect.stack()
+            index = -1
+            for stack_frame in full_stack:
+                print(stack_frame[1],stack_frame[3])
+                #stack_frame[1] -> file name
+                #stack_frame[3] -> method
+                if 'test_' in stack_frame[1]:
+                    index = full_stack.index(stack_frame)
+                    break
+            test_file = full_stack[index][1]
+            test_file = test_file.split(os.sep)[-1]
+            testname = test_file.split('.py')[0]
+            self.set_calling_module(testname)
+
+        return self.calling_module
