+++ core_helpers/screenshot_objects.py
@@ -3,7 +3,6 @@ Helper class for Screenshot Objects
 """
 import os
 import shutil
-from utils import Gif_Maker
 import conf.screenshot_conf as conf
 
 class Screenshot_Objects:
@@ -51,12 +50,6 @@ class Screenshot_Objects:
             self.write("Exception when trying to get rplogger")
             self.write(str(e))
             self.exceptions.append("Error when trying to get reportportal logger")
-
-    def make_gif(self):
-        "Create a gif of all the screenshots within the screenshots directory"
-        self.gif_file_name = Gif_Maker.make_gif(self.screenshot_dir,name=self.calling_module)
-
-        return self.gif_file_name
 
     def set_directory_structure(self):
         "Setup the required directory structure if it is not already present"
