+++ core_helpers/web_app_helper.py
@@ -36,7 +36,7 @@ base_url = conf.base_url_conf
 class Web_App_Helper(Borg, Selenium_Action_Objects, Logging_Objects, Remote_Objects, Screenshot_Objects):
     "Page class that all page models can inherit from"
 
-    def __init__(self,base_url,api_test_flag=False):
+    def __init__(self,base_url):
         "Constructor"
         Borg.__init__(self)
         if self.is_first_time():
@@ -53,9 +53,9 @@ class Web_App_Helper(Borg, Selenium_Action_Objects, Logging_Objects, Remote_Obje
             self.browserstack_flag = False
             self.highlight_flag = False
             self.test_run_id = None
+            self.screenshot_dir = None
             self.reset()
         self.base_url = base_url
-        self.api_test_flag = api_test_flag
         self.driver_obj = DriverFactory()
         if self.driver is not None:
             self.start() #Visit and initialize xpaths for the appropriate page
@@ -104,29 +104,6 @@ class Web_App_Helper(Borg, Selenium_Action_Objects, Logging_Objects, Remote_Obje
         "Turn off the highlighting feature"
         self.highlight_flag = False
 
-    def set_calling_module(self,name):
-        "Set the test name"
-        self.calling_module = name
-
-    def get_calling_module(self):
-        "Get the name of the calling module"
-        if self.calling_module is None:
-            #Try to intelligently figure out name of test when not using pytest
-            full_stack = inspect.stack()
-            index = -1
-            for stack_frame in full_stack:
-                print(stack_frame[1],stack_frame[3])
-                #stack_frame[1] -> file name
-                #stack_frame[3] -> method
-                if 'test_' in stack_frame[1]:
-                    index = full_stack.index(stack_frame)
-                    break
-            test_file = full_stack[index][1]
-            test_file = test_file.split(os.sep)[-1]
-            testname = test_file.split('.py')[0]
-            self.set_calling_module(testname)
-
-        return self.calling_module
 
     def set_screenshot_dir(self,os_name,os_version,browser,browser_version):
         "Set the screenshot directory"
@@ -315,22 +292,6 @@ class Web_App_Helper(Borg, Selenium_Action_Objects, Logging_Objects, Remote_Obje
             self.write("Exception when reading Browser Console log")
             self.write(str(e))
             return log
-
-    def conditional_write(self,flag,positive,negative,level='info'):
-        "Write out either the positive or the negative message based on flag"
-        self.mini_check_counter += 1
-        if level.lower() == "inverse":
-            if flag is True:
-                self.write(positive,level='error')
-            else:
-                self.write(negative,level='info')
-                self.mini_check_pass_counter += 1
-        else:
-            if flag is True:
-                self.write(positive,level='info')
-                self.mini_check_pass_counter += 1
-            else:
-                self.write(negative,level='error')
 
     def execute_javascript(self,js_script,*args):
         "Execute javascipt"
